Move sockstat value parsing to its own package
